requirements.txt
Flask
PyMuPDF
python-docx
Pillow
pdfplumber